feat: create Ajax class and move heartbeat function into it
